+++ assembly/nix/build-linux-x86-64-nix.sh
@@ -27,6 +27,7 @@ fi
 
 mkdir -p artifacts/lib
 cp ./result/bin/* artifacts/
+test $? -eq 0 || { echo "No artifacts have been built..."; exit 1; }
 chmod +x artifacts/*
 rm -rf result
 nix-build linux-x86-64-tonlib.nix
